Refactor test_mixed_task_types.py into focused test modules
Split the large test file (25,228 tokens) into 4 focused modules:
- test_mixed_task_path_resolution.py: Path resolution and priority tests
- test_mixed_task_operations.py: CRUD operations and scope filtering
- test_mixed_task_lifecycle.py: Task claiming, completion, and workflows
- test_mixed_task_validation.py: Error handling and security validation

All 10 test functions preserved with maintained execution time and coverage.
Improves test maintainability and organization for cross-system test suite.
