+++ .pre-commit-config.yaml
@@ -37,4 +37,3 @@ repos:
         language: system
         types: [python]
         pass_filenames: false
-        always_run: true
