+++ planning/projects/P-standalone-task-system/epics/E-cross-system-prerequisites/features/F-cross-system-test-suite/tasks-done/20250718_211510-T-refactor-test-mixed-task-types.md
@@ -1,14 +1,15 @@
 ---
 kind: task
 id: T-refactor-test-mixed-task-types
+parent: F-cross-system-test-suite
+status: done
 title: Refactor test_mixed_task_types.py into multiple focused test modules
-status: open
 priority: high
 prerequisites: []
 created: '2025-07-18T20:57:48.966733'
-updated: '2025-07-18T20:57:48.966733'
+updated: '2025-07-18T21:01:15.923218'
 schema_version: '1.1'
-parent: F-cross-system-test-suite
+worktree: null
 ---
 ### Purpose
 Refactor the large `test_mixed_task_types.py` file into multiple focused test modules to improve maintainability and organization before adding new cross-system tests.
@@ -44,3 +45,6 @@ Refactor the large `test_mixed_task_types.py` file into multiple focused test mo
 
 ### Log
 
+
+**2025-07-19T02:15:10.122989Z** - Successfully refactored the large test_mixed_task_types.py file (25,228 tokens) into 4 focused test modules, improving maintainability and organization. All 10 test functions preserved exactly with clear module boundaries for path resolution, operations, lifecycle, and validation concerns. All tests pass with maintained execution time and coverage.
+- filesChanged: ["tests/integration/test_mixed_task_path_resolution.py", "tests/integration/test_mixed_task_operations.py", "tests/integration/test_mixed_task_lifecycle.py", "tests/integration/test_mixed_task_validation.py"]
