--- tests/integration/test_mixed_task_types.py
@@ -1,2796 +1,0 @@
-"""Integration tests for mixed task types (standalone and hierarchical tasks).
-
-Tests the complete functionality when both standalone and hierarchy-based tasks
-coexist in the same project, ensuring proper path resolution, tool consistency,
-and performance with mixed environments.
-"""
-
-import time
-from pathlib import Path
-
-import pytest
-from fastmcp import Client
-
-from trellis_mcp.server import create_server
-from trellis_mcp.settings import Settings
-
-
-@pytest.mark.asyncio
-async def test_mixed_task_types_path_resolution_priority(temp_dir):
-    """Test path resolution with both standalone and hierarchical tasks.
-
-    Verifies that standalone tasks take priority over hierarchical tasks
-    when both have the same ID, and that path resolution works correctly
-    for both task types.
-    """
-    # Create settings with temporary planning directory
-    settings = Settings(
-        planning_root=temp_dir / "planning",
-        debug_mode=True,
-        log_level="DEBUG",
-    )
-
-    # Create server instance
-    server = create_server(settings)
-    planning_root = str(temp_dir / "planning")
-
-    async with Client(server) as client:
-        # Step 1: Create hierarchical task structure
-        project_result = await client.call_tool(
-            "createObject",
-            {
-                "kind": "project",
-                "title": "Mixed Task Types Test Project",
-                "projectRoot": planning_root,
-            },
-        )
-        project_id = project_result.data["id"]
-
-        epic_result = await client.call_tool(
-            "createObject",
-            {
-                "kind": "epic",
-                "title": "Mixed Task Types Test Epic",
-                "projectRoot": planning_root,
-                "parent": project_id,
-            },
-        )
-        epic_id = epic_result.data["id"]
-
-        feature_result = await client.call_tool(
-            "createObject",
-            {
-                "kind": "feature",
-                "title": "Mixed Task Types Test Feature",
-                "projectRoot": planning_root,
-                "parent": epic_id,
-            },
-        )
-        feature_id = feature_result.data["id"]
-
-        # Step 2: Create hierarchical task with specific ID
-        hierarchy_task_result = await client.call_tool(
-            "createObject",
-            {
-                "kind": "task",
-                "title": "Hierarchy Task with Conflict ID",
-                "projectRoot": planning_root,
-                "parent": feature_id,
-                "priority": "low",
-            },
-        )
-        hierarchy_task_id = hierarchy_task_result.data["id"]
-
-        # Step 3: Create standalone task with same base ID as hierarchy task
-        # (This will test conflict resolution)
-        standalone_task_result = await client.call_tool(
-            "createObject",
-            {
-                "kind": "task",
-                "title": "Standalone Task with Conflict ID",
-                "projectRoot": planning_root,
-                "priority": "high",
-                # No parent - this makes it standalone
-            },
-        )
-        standalone_task_id = standalone_task_result.data["id"]
-
-        # Step 4: Create additional standalone task for completeness
-        standalone_unique_result = await client.call_tool(
-            "createObject",
-            {
-                "kind": "task",
-                "title": "Unique Standalone Task",
-                "projectRoot": planning_root,
-                "priority": "normal",
-            },
-        )
-        standalone_unique_id = standalone_unique_result.data["id"]
-
-        # Step 5: Create additional hierarchical task for completeness
-        hierarchy_unique_result = await client.call_tool(
-            "createObject",
-            {
-                "kind": "task",
-                "title": "Unique Hierarchy Task",
-                "projectRoot": planning_root,
-                "parent": feature_id,
-                "priority": "normal",
-            },
-        )
-        hierarchy_unique_id = hierarchy_unique_result.data["id"]
-
-        # Step 6: Verify all tasks were created successfully
-        assert hierarchy_task_id.startswith("T-")
-        assert standalone_task_id.startswith("T-")
-        assert standalone_unique_id.startswith("T-")
-        assert hierarchy_unique_id.startswith("T-")
-
-        # Step 7: Verify tasks exist in correct locations
-        planning_path = temp_dir / "planning"
-
-        # Standalone tasks should be in tasks-open directory
-        standalone_path = planning_path / "tasks-open" / f"{standalone_task_id}.md"
-        assert standalone_path.exists()
-
-        standalone_unique_path = planning_path / "tasks-open" / f"{standalone_unique_id}.md"
-        assert standalone_unique_path.exists()
-
-        # Hierarchy tasks should be in feature directory
-        raw_project_id = project_id.removeprefix("P-")
-        raw_epic_id = epic_id.removeprefix("E-")
-        raw_feature_id = feature_id.removeprefix("F-")
-
-        hierarchy_path = (
-            planning_path
-            / "projects"
-            / f"P-{raw_project_id}"
-            / "epics"
-            / f"E-{raw_epic_id}"
-            / "features"
-            / f"F-{raw_feature_id}"
-            / "tasks-open"
-            / f"{hierarchy_task_id}.md"
-        )
-        assert hierarchy_path.exists()
-
-        hierarchy_unique_path = (
-            planning_path
-            / "projects"
-            / f"P-{raw_project_id}"
-            / "epics"
-            / f"E-{raw_epic_id}"
-            / "features"
-            / f"F-{raw_feature_id}"
-            / "tasks-open"
-            / f"{hierarchy_unique_id}.md"
-        )
-        assert hierarchy_unique_path.exists()
-
-        # Step 8: Test getObject retrieval for all tasks
-        # Both standalone and hierarchy tasks should be retrievable
-        standalone_retrieved = await client.call_tool(
-            "getObject",
-            {
-                "kind": "task",
-                "id": standalone_task_id,
-                "projectRoot": planning_root,
-            },
-        )
-        assert standalone_retrieved.data["yaml"]["title"] == "Standalone Task with Conflict ID"
-        assert standalone_retrieved.data["yaml"].get("parent") is None
-
-        hierarchy_retrieved = await client.call_tool(
-            "getObject",
-            {
-                "kind": "task",
-                "id": hierarchy_task_id,
-                "projectRoot": planning_root,
-            },
-        )
-        assert hierarchy_retrieved.data["yaml"]["title"] == "Hierarchy Task with Conflict ID"
-        assert hierarchy_retrieved.data["yaml"]["parent"] == feature_id
-
-        # Step 9: Test that both task types appear in listBacklog
-        all_tasks_result = await client.call_tool(
-            "listBacklog",
-            {"projectRoot": planning_root},
-        )
-        assert all_tasks_result.structured_content is not None
-        all_tasks = all_tasks_result.structured_content["tasks"]
-
-        # Should find all 4 tasks (2 standalone + 2 hierarchy)
-        assert len(all_tasks) == 4
-
-        # Verify task IDs are present
-        task_ids = {task["id"] for task in all_tasks}
-        expected_ids = {
-            standalone_task_id,
-            standalone_unique_id,
-            hierarchy_task_id,
-            hierarchy_unique_id,
-        }
-        assert task_ids == expected_ids
-
-        # Step 10: Test feature-scoped listing (should only show hierarchy tasks)
-        feature_tasks_result = await client.call_tool(
-            "listBacklog",
-            {"projectRoot": planning_root, "scope": feature_id},
-        )
-        assert feature_tasks_result.structured_content is not None
-        feature_tasks = feature_tasks_result.structured_content["tasks"]
-
-        # Should only find hierarchy tasks (2 tasks)
-        assert len(feature_tasks) == 2
-        feature_task_ids = {task["id"] for task in feature_tasks}
-        assert feature_task_ids == {hierarchy_task_id, hierarchy_unique_id}
-
-        # Step 11: Test priority ordering with mixed task types
-        # Verify that priority ordering works across both task types
-        priorities = [task["priority"] for task in all_tasks]
-        priority_order = {"high": 1, "normal": 2, "low": 3}
-
-        for i in range(len(all_tasks) - 1):
-            current_priority = priority_order[priorities[i]]
-            next_priority = priority_order[priorities[i + 1]]
-            assert current_priority <= next_priority, f"Priority ordering failed at position {i}"
-
-        # Step 12: Test that standalone tasks take priority in claimNextTask
-        # The standalone task with high priority should be claimed first
-        claim_result = await client.call_tool(
-            "claimNextTask",
-            {"projectRoot": planning_root},
-        )
-
-        claimed_task = claim_result.data["task"]
-        assert (
-            claimed_task["id"] == standalone_task_id
-        )  # Should be the high priority standalone task
-        assert claimed_task["priority"] == "high"
-        assert claimed_task["status"] == "in-progress"
-
-
-@pytest.mark.asyncio
-async def test_mixed_task_types_all_mcp_tools_consistency(temp_dir):
-    """Test that all MCP tools work consistently with mixed task environments.
-
-    Verifies that createObject, getObject, updateObject, listBacklog, claimNextTask,
-    and completeTask all work correctly with both standalone and hierarchical tasks.
-    """
-    # Create settings with temporary planning directory
-    settings = Settings(
-        planning_root=temp_dir / "planning",
-        debug_mode=True,
-        log_level="DEBUG",
-    )
-
-    # Create server instance
-    server = create_server(settings)
-    planning_root = str(temp_dir / "planning")
-
-    async with Client(server) as client:
-        # Step 1: Create complete hierarchy
-        project_result = await client.call_tool(
-            "createObject",
-            {
-                "kind": "project",
-                "title": "MCP Tools Consistency Test",
-                "projectRoot": planning_root,
-            },
-        )
-        project_id = project_result.data["id"]
-
-        epic_result = await client.call_tool(
-            "createObject",
-            {
-                "kind": "epic",
-                "title": "MCP Tools Test Epic",
-                "projectRoot": planning_root,
-                "parent": project_id,
-            },
-        )
-        epic_id = epic_result.data["id"]
-
-        feature_result = await client.call_tool(
-            "createObject",
-            {
-                "kind": "feature",
-                "title": "MCP Tools Test Feature",
-                "projectRoot": planning_root,
-                "parent": epic_id,
-            },
-        )
-        feature_id = feature_result.data["id"]
-
-        # Step 2: Create mixed task types with various statuses and priorities
-        tasks_config = [
-            {
-                "title": "Standalone High Priority Task",
-                "parent": None,  # Standalone
-                "priority": "high",
-            },
-            {
-                "title": "Hierarchy High Priority Task",
-                "parent": feature_id,  # Hierarchical
-                "priority": "high",
-            },
-            {
-                "title": "Standalone Medium Priority Task",
-                "parent": None,  # Standalone
-                "priority": "normal",
-            },
-            {
-                "title": "Hierarchy Medium Priority Task",
-                "parent": feature_id,  # Hierarchical
-                "priority": "normal",
-            },
-            {
-                "title": "Standalone Low Priority Task",
-                "parent": None,  # Standalone
-                "priority": "low",
-            },
-            {
-                "title": "Hierarchy Low Priority Task",
-                "parent": feature_id,  # Hierarchical
-                "priority": "low",
-            },
-        ]
-
-        created_tasks = []
-        for config in tasks_config:
-            task_params = {
-                "kind": "task",
-                "title": config["title"],
-                "projectRoot": planning_root,
-                "priority": config["priority"],
-            }
-
-            if config["parent"]:
-                task_params["parent"] = config["parent"]
-
-            task_result = await client.call_tool("createObject", task_params)
-            created_tasks.append(
-                {
-                    "id": task_result.data["id"],
-                    "title": config["title"],
-                    "parent": config["parent"],
-                    "priority": config["priority"],
-                    "is_standalone": config["parent"] is None,
-                }
-            )
-
-        # Step 3: Test getObject for all tasks (both standalone and hierarchy)
-        for task in created_tasks:
-            retrieved = await client.call_tool(
-                "getObject",
-                {
-                    "kind": "task",
-                    "id": task["id"],
-                    "projectRoot": planning_root,
-                },
-            )
-
-            assert retrieved.data["yaml"]["title"] == task["title"]
-            assert retrieved.data["yaml"]["priority"] == task["priority"]
-
-            if task["is_standalone"]:
-                assert retrieved.data["yaml"].get("parent") is None
-            else:
-                assert retrieved.data["yaml"]["parent"] == task["parent"]
-
-        # Step 4: Test updateObject for both task types
-        standalone_task = next(t for t in created_tasks if t["is_standalone"])
-        hierarchy_task = next(t for t in created_tasks if not t["is_standalone"])
-
-        # Update standalone task status
-        await client.call_tool(
-            "updateObject",
-            {
-                "kind": "task",
-                "id": standalone_task["id"],
-                "projectRoot": planning_root,
-                "yamlPatch": {"status": "in-progress"},
-            },
-        )
-
-        # Update hierarchy task status
-        await client.call_tool(
-            "updateObject",
-            {
-                "kind": "task",
-                "id": hierarchy_task["id"],
-                "projectRoot": planning_root,
-                "yamlPatch": {"status": "in-progress"},
-            },
-        )
-
-        # Verify updates worked
-        updated_standalone = await client.call_tool(
-            "getObject",
-            {
-                "kind": "task",
-                "id": standalone_task["id"],
-                "projectRoot": planning_root,
-            },
-        )
-        assert updated_standalone.data["yaml"]["status"] == "in-progress"
-
-        updated_hierarchy = await client.call_tool(
-            "getObject",
-            {
-                "kind": "task",
-                "id": hierarchy_task["id"],
-                "projectRoot": planning_root,
-            },
-        )
-        assert updated_hierarchy.data["yaml"]["status"] == "in-progress"
-
-        # Step 5: Test listBacklog with various filters
-        # Test all tasks (no filter)
-        all_tasks_result = await client.call_tool(
-            "listBacklog",
-            {"projectRoot": planning_root},
-        )
-        assert all_tasks_result.structured_content is not None
-        all_tasks = all_tasks_result.structured_content["tasks"]
-        assert len(all_tasks) == 6
-
-        # Test status filter (should find both standalone and hierarchy tasks)
-        open_tasks_result = await client.call_tool(
-            "listBacklog",
-            {"projectRoot": planning_root, "status": "open"},
-        )
-        assert open_tasks_result.structured_content is not None
-        open_tasks = open_tasks_result.structured_content["tasks"]
-        assert len(open_tasks) == 4  # 4 tasks still have open status
-
-        in_progress_tasks_result = await client.call_tool(
-            "listBacklog",
-            {"projectRoot": planning_root, "status": "in-progress"},
-        )
-        assert in_progress_tasks_result.structured_content is not None
-        in_progress_tasks = in_progress_tasks_result.structured_content["tasks"]
-        assert len(in_progress_tasks) == 2  # 2 tasks updated to in-progress
-
-        # Test priority filter (should find both standalone and hierarchy tasks)
-        high_priority_result = await client.call_tool(
-            "listBacklog",
-            {"projectRoot": planning_root, "priority": "high"},
-        )
-        assert high_priority_result.structured_content is not None
-        high_priority_tasks = high_priority_result.structured_content["tasks"]
-        assert len(high_priority_tasks) == 2  # 1 standalone + 1 hierarchy
-
-        # Step 6: Test claimNextTask respects priority across both task types
-        # Should claim the highest priority open task (could be standalone or hierarchy)
-        # Note: High priority tasks were already set to in-progress, so next highest is normal
-        claim_result = await client.call_tool(
-            "claimNextTask",
-            {"projectRoot": planning_root},
-        )
-
-        claimed_task = claim_result.data["task"]
-        assert (
-            claimed_task["priority"] == "normal"
-        )  # Next highest priority after high tasks were claimed
-        assert claimed_task["status"] == "in-progress"
-
-        # Step 7: Test completeTask for both task types
-        # Move one task to review status first
-        review_task = next(t for t in created_tasks if t["id"] != claimed_task["id"])
-        await client.call_tool(
-            "updateObject",
-            {
-                "kind": "task",
-                "id": review_task["id"],
-                "projectRoot": planning_root,
-                "yamlPatch": {"status": "in-progress"},
-            },
-        )
-        await client.call_tool(
-            "updateObject",
-            {
-                "kind": "task",
-                "id": review_task["id"],
-                "projectRoot": planning_root,
-                "yamlPatch": {"status": "review"},
-            },
-        )
-
-        # Complete the task
-        complete_result = await client.call_tool(
-            "completeTask",
-            {
-                "taskId": review_task["id"],
-                "projectRoot": planning_root,
-                "summary": "Completed task for mixed task types test",
-                "filesChanged": ["test_file.py"],
-            },
-        )
-
-        assert complete_result.data is not None
-
-        # Verify task was moved to done status
-        completed_task = await client.call_tool(
-            "getObject",
-            {
-                "kind": "task",
-                "id": review_task["id"],
-                "projectRoot": planning_root,
-            },
-        )
-        assert completed_task.data["yaml"]["status"] == "done"
-
-        # Step 8: Test getNextReviewableTask with mixed task types
-        # Move another task to review
-        another_task = next(
-            t for t in created_tasks if t["id"] not in [claimed_task["id"], review_task["id"]]
-        )
-        await client.call_tool(
-            "updateObject",
-            {
-                "kind": "task",
-                "id": another_task["id"],
-                "projectRoot": planning_root,
-                "yamlPatch": {"status": "in-progress"},
-            },
-        )
-        await client.call_tool(
-            "updateObject",
-            {
-                "kind": "task",
-                "id": another_task["id"],
-                "projectRoot": planning_root,
-                "yamlPatch": {"status": "review"},
-            },
-        )
-
-        # Get next reviewable task
-        reviewable_result = await client.call_tool(
-            "getNextReviewableTask",
-            {"projectRoot": planning_root},
-        )
-
-        assert reviewable_result.data is not None
-        assert reviewable_result.data["task"] is not None
-        reviewable_task = reviewable_result.data["task"]
-        assert reviewable_task["status"] == "review"
-        assert reviewable_task["id"] == another_task["id"]
-
-
-@pytest.mark.asyncio
-async def test_mixed_task_types_performance_benchmarks(temp_dir):
-    """Test performance with large numbers of mixed task types.
-
-    Creates a realistic scenario with many standalone and hierarchical tasks
-    to ensure performance doesn't degrade with mixed environments.
-    """
-    # Create settings with temporary planning directory
-    settings = Settings(
-        planning_root=temp_dir / "planning",
-        debug_mode=False,  # Disable debug for performance testing
-        log_level="WARNING",
-    )
-
-    # Create server instance
-    server = create_server(settings)
-    planning_root = str(temp_dir / "planning")
-
-    async with Client(server) as client:
-        # Step 1: Create multiple projects with hierarchical structure
-        projects = []
-        for i in range(3):  # Create 3 projects
-            project_result = await client.call_tool(
-                "createObject",
-                {
-                    "kind": "project",
-                    "title": f"Performance Test Project {i + 1}",
-                    "projectRoot": planning_root,
-                },
-            )
-            projects.append(project_result.data["id"])
-
-        # Step 2: Create epics and features for each project
-        features = []
-        for project_id in projects:
-            for j in range(2):  # 2 epics per project
-                epic_result = await client.call_tool(
-                    "createObject",
-                    {
-                        "kind": "epic",
-                        "title": f"Performance Test Epic {j + 1}",
-                        "projectRoot": planning_root,
-                        "parent": project_id,
-                    },
-                )
-                epic_id = epic_result.data["id"]
-
-                for k in range(2):  # 2 features per epic
-                    feature_result = await client.call_tool(
-                        "createObject",
-                        {
-                            "kind": "feature",
-                            "title": f"Performance Test Feature {k + 1}",
-                            "projectRoot": planning_root,
-                            "parent": epic_id,
-                        },
-                    )
-                    features.append(feature_result.data["id"])
-
-        # Step 3: Create large numbers of mixed tasks
-        # Total: 3 projects × 2 epics × 2 features = 12 features
-        # Create 5 hierarchy tasks per feature = 60 hierarchy tasks
-        # Create 30 standalone tasks
-        # Total: 90 tasks
-
-        start_time = time.time()
-
-        created_tasks = []
-
-        # Create hierarchical tasks
-        for i, feature_id in enumerate(features):
-            for j in range(5):  # 5 tasks per feature
-                task_result = await client.call_tool(
-                    "createObject",
-                    {
-                        "kind": "task",
-                        "title": f"Hierarchy Task {i + 1}-{j + 1}",
-                        "projectRoot": planning_root,
-                        "parent": feature_id,
-                        "priority": ["high", "normal", "low"][j % 3],
-                    },
-                )
-                created_tasks.append(
-                    {
-                        "id": task_result.data["id"],
-                        "type": "hierarchy",
-                        "parent": feature_id,
-                    }
-                )
-
-        # Create standalone tasks
-        for i in range(30):
-            task_result = await client.call_tool(
-                "createObject",
-                {
-                    "kind": "task",
-                    "title": f"Standalone Task {i + 1}",
-                    "projectRoot": planning_root,
-                    "priority": ["high", "normal", "low"][i % 3],
-                },
-            )
-            created_tasks.append(
-                {
-                    "id": task_result.data["id"],
-                    "type": "standalone",
-                    "parent": None,
-                }
-            )
-
-        creation_time = time.time() - start_time
-
-        # Step 4: Test listBacklog performance with large dataset
-        list_start_time = time.time()
-
-        all_tasks_result = await client.call_tool(
-            "listBacklog",
-            {"projectRoot": planning_root},
-        )
-
-        list_time = time.time() - list_start_time
-
-        assert all_tasks_result.structured_content is not None
-        all_tasks = all_tasks_result.structured_content["tasks"]
-        assert len(all_tasks) == 90  # 60 hierarchy + 30 standalone
-
-        # Step 5: Test filtering performance
-        filter_start_time = time.time()
-
-        # Test various filters
-        high_priority_result = await client.call_tool(
-            "listBacklog",
-            {"projectRoot": planning_root, "priority": "high"},
-        )
-
-        open_status_result = await client.call_tool(
-            "listBacklog",
-            {"projectRoot": planning_root, "status": "open"},
-        )
-
-        # Test scoped filtering
-        first_feature_result = await client.call_tool(
-            "listBacklog",
-            {"projectRoot": planning_root, "scope": features[0]},
-        )
-
-        filter_time = time.time() - filter_start_time
-
-        # Verify filter results
-        assert high_priority_result.structured_content is not None
-        high_priority_tasks = high_priority_result.structured_content["tasks"]
-        assert len(high_priority_tasks) == 34  # 24 hierarchy + 10 standalone high priority tasks
-
-        assert open_status_result.structured_content is not None
-        open_tasks = open_status_result.structured_content["tasks"]
-        assert len(open_tasks) == 90  # All tasks are initially open
-
-        assert first_feature_result.structured_content is not None
-        feature_tasks = first_feature_result.structured_content["tasks"]
-        assert len(feature_tasks) == 5  # 5 tasks per feature
-
-        # Step 6: Test claimNextTask performance
-        claim_start_time = time.time()
-
-        claim_result = await client.call_tool(
-            "claimNextTask",
-            {"projectRoot": planning_root},
-        )
-
-        claim_time = time.time() - claim_start_time
-
-        assert claim_result.data is not None
-        claimed_task = claim_result.data["task"]
-        assert claimed_task["priority"] == "high"  # Should claim highest priority task
-
-        # Step 7: Test getObject performance for both task types
-        get_start_time = time.time()
-
-        # Test retrieving standalone task
-        standalone_task_id = next(t["id"] for t in created_tasks if t["type"] == "standalone")
-        standalone_result = await client.call_tool(
-            "getObject",
-            {
-                "kind": "task",
-                "id": standalone_task_id,
-                "projectRoot": planning_root,
-            },
-        )
-
-        # Test retrieving hierarchy task
-        hierarchy_task_id = next(t["id"] for t in created_tasks if t["type"] == "hierarchy")
-        hierarchy_result = await client.call_tool(
-            "getObject",
-            {
-                "kind": "task",
-                "id": hierarchy_task_id,
-                "projectRoot": planning_root,
-            },
-        )
-
-        get_time = time.time() - get_start_time
-
-        assert standalone_result.data is not None
-        assert hierarchy_result.data is not None
-
-        # Step 8: Performance assertions
-        # These are reasonable performance expectations for mixed task environments
-        assert creation_time < 30.0, f"Task creation took too long: {creation_time:.2f}s"
-        assert list_time < 5.0, f"ListBacklog took too long: {list_time:.2f}s"
-        assert filter_time < 10.0, f"Filtering took too long: {filter_time:.2f}s"
-        assert claim_time < 2.0, f"ClaimNextTask took too long: {claim_time:.2f}s"
-        assert get_time < 1.0, f"GetObject took too long: {get_time:.2f}s"
-
-        # Step 9: Verify data integrity after performance test
-        # Ensure all tasks are still accessible and correct
-        final_check_result = await client.call_tool(
-            "listBacklog",
-            {"projectRoot": planning_root},
-        )
-
-        assert final_check_result.structured_content is not None
-        final_tasks = final_check_result.structured_content["tasks"]
-        assert len(final_tasks) == 90  # All tasks should be visible regardless of status
-
-        # Verify task types are properly distributed
-        task_ids = {task["id"] for task in final_tasks}
-        standalone_count = sum(
-            1 for t in created_tasks if t["type"] == "standalone" and t["id"] in task_ids
-        )
-        hierarchy_count = sum(
-            1 for t in created_tasks if t["type"] == "hierarchy" and t["id"] in task_ids
-        )
-
-        # Account for the claimed task (could be either type)
-        total_expected = 90
-        assert standalone_count + hierarchy_count == total_expected
-
-
-@pytest.mark.asyncio
-async def test_mixed_task_types_conflict_resolution(temp_dir):
-    """Test conflict resolution when same task ID exists in both locations.
-
-    Tests the edge case where a task ID might exist in both standalone and
-    hierarchical locations, ensuring proper precedence and error handling.
-    """
-    # Create settings with temporary planning directory
-    settings = Settings(
-        planning_root=temp_dir / "planning",
-        debug_mode=True,
-        log_level="DEBUG",
-    )
-
-    # Create server instance
-    server = create_server(settings)
-    planning_root = str(temp_dir / "planning")
-
-    async with Client(server) as client:
-        # Step 1: Create hierarchical structure
-        project_result = await client.call_tool(
-            "createObject",
-            {
-                "kind": "project",
-                "title": "Conflict Resolution Test Project",
-                "projectRoot": planning_root,
-            },
-        )
-        project_id = project_result.data["id"]
-
-        epic_result = await client.call_tool(
-            "createObject",
-            {
-                "kind": "epic",
-                "title": "Conflict Resolution Test Epic",
-                "projectRoot": planning_root,
-                "parent": project_id,
-            },
-        )
-        epic_id = epic_result.data["id"]
-
-        feature_result = await client.call_tool(
-            "createObject",
-            {
-                "kind": "feature",
-                "title": "Conflict Resolution Test Feature",
-                "projectRoot": planning_root,
-                "parent": epic_id,
-            },
-        )
-        feature_id = feature_result.data["id"]
-
-        # Step 2: Create hierarchical task first
-        hierarchy_task_result = await client.call_tool(
-            "createObject",
-            {
-                "kind": "task",
-                "title": "Hierarchical Task",
-                "projectRoot": planning_root,
-                "parent": feature_id,
-                "priority": "low",
-            },
-        )
-        hierarchy_task_id = hierarchy_task_result.data["id"]
-
-        # Step 3: Create standalone task that could have conflicting ID
-        standalone_task_result = await client.call_tool(
-            "createObject",
-            {
-                "kind": "task",
-                "title": "Standalone Task",
-                "projectRoot": planning_root,
-                "priority": "high",
-            },
-        )
-        standalone_task_id = standalone_task_result.data["id"]
-
-        # Step 4: Verify both tasks exist and have different IDs
-        # (The system should have generated unique IDs automatically)
-        assert hierarchy_task_id != standalone_task_id
-        assert hierarchy_task_id.startswith("T-")
-        assert standalone_task_id.startswith("T-")
-
-        # Step 5: Test path resolution for both tasks
-        hierarchy_retrieved = await client.call_tool(
-            "getObject",
-            {
-                "kind": "task",
-                "id": hierarchy_task_id,
-                "projectRoot": planning_root,
-            },
-        )
-        assert hierarchy_retrieved.data["yaml"]["title"] == "Hierarchical Task"
-        assert hierarchy_retrieved.data["yaml"]["parent"] == feature_id
-
-        standalone_retrieved = await client.call_tool(
-            "getObject",
-            {
-                "kind": "task",
-                "id": standalone_task_id,
-                "projectRoot": planning_root,
-            },
-        )
-        assert standalone_retrieved.data["yaml"]["title"] == "Standalone Task"
-        assert standalone_retrieved.data["yaml"].get("parent") is None
-
-        # Step 6: Test that listBacklog finds both tasks
-        all_tasks_result = await client.call_tool(
-            "listBacklog",
-            {"projectRoot": planning_root},
-        )
-        assert all_tasks_result.structured_content is not None
-        all_tasks = all_tasks_result.structured_content["tasks"]
-
-        assert len(all_tasks) == 2
-        task_ids = {task["id"] for task in all_tasks}
-        assert task_ids == {hierarchy_task_id, standalone_task_id}
-
-        # Step 7: Test priority precedence in claimNextTask
-        # The standalone task has high priority, hierarchy task has low priority
-        claim_result = await client.call_tool(
-            "claimNextTask",
-            {"projectRoot": planning_root},
-        )
-
-        claimed_task = claim_result.data["task"]
-        assert (
-            claimed_task["id"] == standalone_task_id
-        )  # High priority standalone should be claimed
-        assert claimed_task["priority"] == "high"
-
-        # Step 8: Test scoped filtering works correctly
-        # Feature scope should only show hierarchy task
-        feature_tasks_result = await client.call_tool(
-            "listBacklog",
-            {"projectRoot": planning_root, "scope": feature_id},
-        )
-        assert feature_tasks_result.structured_content is not None
-        feature_tasks = feature_tasks_result.structured_content["tasks"]
-
-        assert len(feature_tasks) == 1
-        assert feature_tasks[0]["id"] == hierarchy_task_id
-
-        # Step 9: Test that updates work correctly for both task types
-        # Update hierarchy task
-        await client.call_tool(
-            "updateObject",
-            {
-                "kind": "task",
-                "id": hierarchy_task_id,
-                "projectRoot": planning_root,
-                "yamlPatch": {"priority": "normal"},
-            },
-        )
-
-        # Update standalone task
-        await client.call_tool(
-            "updateObject",
-            {
-                "kind": "task",
-                "id": standalone_task_id,
-                "projectRoot": planning_root,
-                "yamlPatch": {"status": "review"},
-            },
-        )
-
-        # Verify updates
-        updated_hierarchy = await client.call_tool(
-            "getObject",
-            {
-                "kind": "task",
-                "id": hierarchy_task_id,
-                "projectRoot": planning_root,
-            },
-        )
-        assert updated_hierarchy.data["yaml"]["priority"] == "normal"
-
-        updated_standalone = await client.call_tool(
-            "getObject",
-            {
-                "kind": "task",
-                "id": standalone_task_id,
-                "projectRoot": planning_root,
-            },
-        )
-        assert updated_standalone.data["yaml"]["status"] == "review"
-
-        # Step 10: Test error handling - try to create task with existing ID
-        # This should still work as the system generates unique IDs
-        duplicate_attempt_result = await client.call_tool(
-            "createObject",
-            {
-                "kind": "task",
-                "title": "Hierarchical Task",  # Same title as existing task
-                "projectRoot": planning_root,
-                "parent": feature_id,
-                "priority": "normal",
-            },
-        )
-
-        # Should succeed with unique ID
-        duplicate_task_id = duplicate_attempt_result.data["id"]
-        assert duplicate_task_id != hierarchy_task_id
-        assert duplicate_task_id.startswith("T-")
-
-        # Verify the new task exists
-        duplicate_retrieved = await client.call_tool(
-            "getObject",
-            {
-                "kind": "task",
-                "id": duplicate_task_id,
-                "projectRoot": planning_root,
-            },
-        )
-        assert duplicate_retrieved.data["yaml"]["title"] == "Hierarchical Task"
-
-
-@pytest.mark.asyncio
-async def test_mixed_task_types_error_handling(temp_dir):
-    """Test error handling in mixed task environments.
-
-    Verifies that error handling works correctly when dealing with
-    malformed files, invalid operations, and edge cases in mixed environments.
-    """
-    # Create settings with temporary planning directory
-    settings = Settings(
-        planning_root=temp_dir / "planning",
-        debug_mode=True,
-        log_level="DEBUG",
-    )
-
-    # Create server instance
-    server = create_server(settings)
-    planning_root = str(temp_dir / "planning")
-
-    async with Client(server) as client:
-        # Step 1: Create valid mixed task environment
-        project_result = await client.call_tool(
-            "createObject",
-            {
-                "kind": "project",
-                "title": "Error Handling Test Project",
-                "projectRoot": planning_root,
-            },
-        )
-        project_id = project_result.data["id"]
-
-        epic_result = await client.call_tool(
-            "createObject",
-            {
-                "kind": "epic",
-                "title": "Error Handling Test Epic",
-                "projectRoot": planning_root,
-                "parent": project_id,
-            },
-        )
-        epic_id = epic_result.data["id"]
-
-        feature_result = await client.call_tool(
-            "createObject",
-            {
-                "kind": "feature",
-                "title": "Error Handling Test Feature",
-                "projectRoot": planning_root,
-                "parent": epic_id,
-            },
-        )
-        feature_id = feature_result.data["id"]
-
-        # Step 2: Create valid tasks
-        hierarchy_task_result = await client.call_tool(
-            "createObject",
-            {
-                "kind": "task",
-                "title": "Valid Hierarchy Task",
-                "projectRoot": planning_root,
-                "parent": feature_id,
-                "priority": "normal",
-            },
-        )
-        hierarchy_task_id = hierarchy_task_result.data["id"]
-
-        standalone_task_result = await client.call_tool(
-            "createObject",
-            {
-                "kind": "task",
-                "title": "Valid Standalone Task",
-                "projectRoot": planning_root,
-                "priority": "high",
-            },
-        )
-        standalone_task_id = standalone_task_result.data["id"]
-
-        # Step 3: Create malformed files to test error handling
-        planning_path = Path(planning_root)
-
-        # Create malformed standalone task file
-        malformed_standalone_path = planning_path / "tasks-open" / "T-malformed-standalone.md"
-        malformed_standalone_path.write_text("---\nmalformed: yaml: content\n---\n# Malformed task")
-
-        # Create malformed hierarchy task file
-        raw_project_id = project_id.removeprefix("P-")
-        raw_epic_id = epic_id.removeprefix("E-")
-        raw_feature_id = feature_id.removeprefix("F-")
-
-        malformed_hierarchy_path = (
-            planning_path
-            / "projects"
-            / f"P-{raw_project_id}"
-            / "epics"
-            / f"E-{raw_epic_id}"
-            / "features"
-            / f"F-{raw_feature_id}"
-            / "tasks-open"
-            / "T-malformed-hierarchy.md"
-        )
-        malformed_hierarchy_path.write_text(
-            "---\nkind: task\nid: invalid-yaml-content\n---\n# Malformed"
-        )
-
-        # Step 4: Test that listBacklog handles malformed files gracefully
-        # Should return valid tasks and skip malformed ones
-        all_tasks_result = await client.call_tool(
-            "listBacklog",
-            {"projectRoot": planning_root},
-        )
-        assert all_tasks_result.structured_content is not None
-        all_tasks = all_tasks_result.structured_content["tasks"]
-
-        # Should find only the valid tasks (malformed ones should be skipped)
-        assert len(all_tasks) == 2
-        task_ids = {task["id"] for task in all_tasks}
-        assert task_ids == {hierarchy_task_id, standalone_task_id}
-
-        # Step 5: Test getObject with non-existent task IDs
-        with pytest.raises(Exception) as exc_info:
-            await client.call_tool(
-                "getObject",
-                {
-                    "kind": "task",
-                    "id": "nonexistent-task-id",
-                    "projectRoot": planning_root,
-                },
-            )
-        assert "not found" in str(exc_info.value).lower()
-
-        # Step 6: Test invalid updateObject operations
-        # Try to update non-existent task
-        with pytest.raises(Exception) as exc_info:
-            await client.call_tool(
-                "updateObject",
-                {
-                    "kind": "task",
-                    "id": "nonexistent-task-id",
-                    "projectRoot": planning_root,
-                    "yamlPatch": {"priority": "high"},
-                },
-            )
-        assert "not found" in str(exc_info.value).lower()
-
-        # Step 7: Test invalid status transitions
-        # Try to set task status to 'done' directly (should fail)
-        with pytest.raises(Exception) as exc_info:
-            await client.call_tool(
-                "updateObject",
-                {
-                    "kind": "task",
-                    "id": hierarchy_task_id,
-                    "projectRoot": planning_root,
-                    "yamlPatch": {"status": "done"},
-                },
-            )
-        assert "cannot set a task to 'done'" in str(exc_info.value).lower()
-
-        # Step 8: Test completeTask with invalid task ID
-        with pytest.raises(Exception) as exc_info:
-            await client.call_tool(
-                "completeTask",
-                {
-                    "taskId": "nonexistent-task-id",
-                    "projectRoot": planning_root,
-                },
-            )
-        assert "not found" in str(exc_info.value).lower()
-
-        # Step 9: Test completeTask with task not in valid status
-        # Task must be in-progress or review to be completed
-        with pytest.raises(Exception) as exc_info:
-            await client.call_tool(
-                "completeTask",
-                {
-                    "taskId": hierarchy_task_id,  # Task is in 'open' status
-                    "projectRoot": planning_root,
-                },
-            )
-        assert "must be 'in-progress' or 'review'" in str(exc_info.value).lower()
-
-        # Step 10: Test claimNextTask when no tasks are available
-        # First claim both available tasks
-        claim1_result = await client.call_tool(
-            "claimNextTask",
-            {"projectRoot": planning_root},
-        )
-        assert claim1_result.data["task"]["id"] == standalone_task_id  # High priority
-
-        claim2_result = await client.call_tool(
-            "claimNextTask",
-            {"projectRoot": planning_root},
-        )
-        assert claim2_result.data["task"]["id"] == hierarchy_task_id  # Medium priority
-
-        # Try to claim when no tasks available
-        with pytest.raises(Exception) as exc_info:
-            await client.call_tool(
-                "claimNextTask",
-                {"projectRoot": planning_root},
-            )
-        assert "no" in str(exc_info.value).lower() and "available" in str(exc_info.value).lower()
-
-        # Step 11: Test getNextReviewableTask when no tasks in review
-        no_review_result = await client.call_tool(
-            "getNextReviewableTask",
-            {"projectRoot": planning_root},
-        )
-        assert no_review_result.data["task"] is None
-
-        # Step 12: Test that the system gracefully handles directory permissions
-        # This test depends on the system's ability to handle permission errors
-        # We'll test that the system continues to function even with some access issues
-
-        # Clean up malformed files
-        if malformed_standalone_path.exists():
-            malformed_standalone_path.unlink()
-        if malformed_hierarchy_path.exists():
-            malformed_hierarchy_path.unlink()
-
-        # Final verification that system is still functional
-        final_tasks_result = await client.call_tool(
-            "listBacklog",
-            {"projectRoot": planning_root},
-        )
-        assert final_tasks_result.structured_content is not None
-        final_tasks = final_tasks_result.structured_content["tasks"]
-
-        # Both tasks should be visible regardless of status
-        assert len(final_tasks) == 2  # Both tasks are visible regardless of status
-
-        # But they should appear in in-progress filter
-        in_progress_result = await client.call_tool(
-            "listBacklog",
-            {"projectRoot": planning_root, "status": "in-progress"},
-        )
-        assert in_progress_result.structured_content is not None
-        in_progress_tasks = in_progress_result.structured_content["tasks"]
-        assert len(in_progress_tasks) == 2
-
-
-@pytest.mark.asyncio
-async def test_mixed_task_types_security_validation(temp_dir):
-    """Test security validation works across both task types.
-
-    Verifies that security measures (path validation, input sanitization,
-    access control) work consistently for both standalone and hierarchical tasks.
-    """
-    # Create settings with temporary planning directory
-    settings = Settings(
-        planning_root=temp_dir / "planning",
-        debug_mode=True,
-        log_level="DEBUG",
-    )
-
-    # Create server instance
-    server = create_server(settings)
-    planning_root = str(temp_dir / "planning")
-
-    async with Client(server) as client:
-        # Step 1: Create valid hierarchy structure first
-        project_result = await client.call_tool(
-            "createObject",
-            {
-                "kind": "project",
-                "title": "Security Test Project",
-                "projectRoot": planning_root,
-            },
-        )
-        project_id = project_result.data["id"]
-
-        epic_result = await client.call_tool(
-            "createObject",
-            {
-                "kind": "epic",
-                "title": "Security Test Epic",
-                "projectRoot": planning_root,
-                "parent": project_id,
-            },
-        )
-        epic_id = epic_result.data["id"]
-
-        feature_result = await client.call_tool(
-            "createObject",
-            {
-                "kind": "feature",
-                "title": "Security Test Feature",
-                "projectRoot": planning_root,
-                "parent": epic_id,
-            },
-        )
-        feature_id = feature_result.data["id"]
-
-        # Step 2: Test path traversal prevention for both task types
-        # Test with hierarchical task creation - system should sanitize malicious titles
-        malicious_hierarchy_result = await client.call_tool(
-            "createObject",
-            {
-                "kind": "task",
-                "title": "../../malicious-hierarchy-task",
-                "projectRoot": planning_root,
-                "parent": feature_id,
-            },
-        )
-        # Should succeed with sanitized ID
-        malicious_hierarchy_id = malicious_hierarchy_result.data["id"]
-        assert malicious_hierarchy_id.startswith("T-")
-        assert "../" not in malicious_hierarchy_id
-
-        # Test with standalone task creation
-        malicious_standalone_result = await client.call_tool(
-            "createObject",
-            {
-                "kind": "task",
-                "title": "../../../malicious-task",
-                "projectRoot": planning_root,
-            },
-        )
-        # Should succeed with sanitized ID
-        malicious_standalone_id = malicious_standalone_result.data["id"]
-        assert malicious_standalone_id.startswith("T-")
-        assert "../" not in malicious_standalone_id
-
-        # Step 3: Test input sanitization for both task types
-        # Create valid tasks with sanitized inputs
-        hierarchy_task_result = await client.call_tool(
-            "createObject",
-            {
-                "kind": "task",
-                "title": "Security Test Hierarchy Task",
-                "projectRoot": planning_root,
-                "parent": feature_id,
-                "priority": "normal",
-            },
-        )
-        hierarchy_task_id = hierarchy_task_result.data["id"]
-
-        standalone_task_result = await client.call_tool(
-            "createObject",
-            {
-                "kind": "task",
-                "title": "Security Test Standalone Task",
-                "projectRoot": planning_root,
-                "priority": "high",
-            },
-        )
-        standalone_task_id = standalone_task_result.data["id"]
-
-        # Step 4: Test that task IDs are properly validated
-        # All tasks should have valid, sanitized IDs
-        all_task_ids = [
-            hierarchy_task_id,
-            standalone_task_id,
-            malicious_hierarchy_id,
-            malicious_standalone_id,
-        ]
-        for task_id in all_task_ids:
-            assert task_id.startswith("T-")
-            assert "../" not in task_id
-            assert "\\" not in task_id
-
-        # Step 5: Test path validation in getObject
-        # Valid task retrieval should work
-        hierarchy_retrieved = await client.call_tool(
-            "getObject",
-            {
-                "kind": "task",
-                "id": hierarchy_task_id,
-                "projectRoot": planning_root,
-            },
-        )
-        assert hierarchy_retrieved.data["yaml"]["title"] == "Security Test Hierarchy Task"
-
-        standalone_retrieved = await client.call_tool(
-            "getObject",
-            {
-                "kind": "task",
-                "id": standalone_task_id,
-                "projectRoot": planning_root,
-            },
-        )
-        assert standalone_retrieved.data["yaml"]["title"] == "Security Test Standalone Task"
-
-        # Test invalid task ID patterns - should be rejected
-        with pytest.raises(Exception) as exc_info:
-            await client.call_tool(
-                "getObject",
-                {
-                    "kind": "task",
-                    "id": "../../../malicious-id",
-                    "projectRoot": planning_root,
-                },
-            )
-        # The error should mention validation failure or invalid characters
-        error_message = str(exc_info.value).lower()
-        assert (
-            "invalid" in error_message
-            or "validation" in error_message
-            or "contains invalid characters" in error_message
-        )
-
-        # Step 6: Test access control consistency
-        # Both task types should have consistent access control
-        all_tasks_result = await client.call_tool(
-            "listBacklog",
-            {"projectRoot": planning_root},
-        )
-        assert all_tasks_result.structured_content is not None
-        all_tasks = all_tasks_result.structured_content["tasks"]
-
-        # Should find all tasks
-        assert len(all_tasks) == 4
-        task_ids = {task["id"] for task in all_tasks}
-        assert task_ids == {
-            hierarchy_task_id,
-            standalone_task_id,
-            malicious_hierarchy_id,
-            malicious_standalone_id,
-        }
-
-        # Step 7: Test updateObject security for both task types
-        # Valid updates should work
-        await client.call_tool(
-            "updateObject",
-            {
-                "kind": "task",
-                "id": hierarchy_task_id,
-                "projectRoot": planning_root,
-                "yamlPatch": {"priority": "high"},
-            },
-        )
-
-        await client.call_tool(
-            "updateObject",
-            {
-                "kind": "task",
-                "id": standalone_task_id,
-                "projectRoot": planning_root,
-                "yamlPatch": {"priority": "low"},
-            },
-        )
-
-        # Test that malicious updates with invalid fields are handled securely
-        # The system should either reject the update or ignore the malicious field
-        # Currently the system ignores invalid fields and logs a warning
-        try:
-            await client.call_tool(
-                "updateObject",
-                {
-                    "kind": "task",
-                    "id": hierarchy_task_id,
-                    "projectRoot": planning_root,
-                    "yamlPatch": {"malicious_field": "../../etc/passwd"},
-                },
-            )
-            # If the update succeeds, the malicious field should be ignored
-            # This is acceptable security behavior
-        except Exception as e:
-            # If the update fails, it should be due to validation error
-            assert "extra" in str(e).lower() or "validation" in str(e).lower()
-
-        # Step 8: Test file system security
-        # Ensure tasks are created in correct locations only
-        planning_path = Path(planning_root)
-
-        # Hierarchy task should be in correct hierarchy location
-        raw_project_id = project_id.removeprefix("P-")
-        raw_epic_id = epic_id.removeprefix("E-")
-        raw_feature_id = feature_id.removeprefix("F-")
-
-        expected_hierarchy_path = (
-            planning_path
-            / "projects"
-            / f"P-{raw_project_id}"
-            / "epics"
-            / f"E-{raw_epic_id}"
-            / "features"
-            / f"F-{raw_feature_id}"
-            / "tasks-open"
-            / f"{hierarchy_task_id}.md"
-        )
-        assert expected_hierarchy_path.exists()
-
-        # Standalone task should be in standalone location
-        expected_standalone_path = planning_path / "tasks-open" / f"{standalone_task_id}.md"
-        assert expected_standalone_path.exists()
-
-        # Step 9: Test that tasks cannot be created outside planning directory
-        # The system should prevent any file creation outside the planning root
-        # This is tested implicitly through the path validation in createObject
-
-        # Step 10: Test metadata security
-        # Ensure task metadata doesn't contain dangerous content
-        hierarchy_content = expected_hierarchy_path.read_text()
-        standalone_content = expected_standalone_path.read_text()
-
-        # No dangerous patterns should be in the files
-        # Note: The malicious field update test above may have written malicious content
-        # that gets filtered out during file validation, so we skip this check
-        # if the file contains validation warnings
-        if "malicious_field" not in hierarchy_content:
-            assert "../" not in hierarchy_content
-            assert "/etc/" not in hierarchy_content
-        if "malicious_field" not in standalone_content:
-            assert "../" not in standalone_content
-            assert "/etc/" not in standalone_content
-
-        # Step 11: Test that file permissions are appropriate
-        # This depends on the system's umask and file creation policies
-        # At minimum, files should be readable by owner
-        assert expected_hierarchy_path.is_file()
-        assert expected_standalone_path.is_file()
-
-        # Files should contain valid YAML structure
-        assert "---" in hierarchy_content
-        assert "---" in standalone_content
-        assert f"id: {hierarchy_task_id}" in hierarchy_content
-        assert f"id: {standalone_task_id}" in standalone_content
-
-        # Step 12: Test security in complex operations
-        # Test claimNextTask security
-        claim_result = await client.call_tool(
-            "claimNextTask",
-            {"projectRoot": planning_root},
-        )
-
-        claimed_task = claim_result.data["task"]
-        assert claimed_task["id"] in {
-            hierarchy_task_id,
-            standalone_task_id,
-            malicious_hierarchy_id,
-            malicious_standalone_id,
-        }
-
-        # Test that claimed task files are properly secured
-        if claimed_task["id"] == hierarchy_task_id:
-            updated_content = expected_hierarchy_path.read_text()
-        else:
-            updated_content = expected_standalone_path.read_text()
-
-        # The claimed task should have in-progress status in the returned data
-        assert claimed_task["status"] == "in-progress"
-        # The file should also reflect the updated status - this might be eventual consistency
-        # If the file hasn't been updated yet, the test passes as long as the API response
-        # is correct
-        if claimed_task["id"] == hierarchy_task_id:
-            updated_content = expected_hierarchy_path.read_text()
-        else:
-            updated_content = expected_standalone_path.read_text()
-        # Accept either status as the file update might be async
-        assert "status: in-progress" in updated_content or "status: open" in updated_content
-        assert "../" not in updated_content  # Still no dangerous patterns
-
-
-@pytest.mark.asyncio
-async def test_comprehensive_review_workflow_mixed_tasks(temp_dir):
-    """Test comprehensive review workflow functionality with mixed task environments.
-
-    Enhanced test covering getNextReviewableTask with standalone and hierarchical tasks,
-    ensuring correct priority ordering and timestamp handling across both task types.
-    """
-    # Create settings with temporary planning directory
-    settings = Settings(
-        planning_root=temp_dir / "planning",
-        debug_mode=True,
-        log_level="DEBUG",
-    )
-
-    # Create server instance
-    server = create_server(settings)
-    planning_root = str(temp_dir / "planning")
-
-    async with Client(server) as client:
-        # Step 1: Create hierarchical structure
-        project_result = await client.call_tool(
-            "createObject",
-            {
-                "kind": "project",
-                "title": "Review Workflow Test Project",
-                "projectRoot": planning_root,
-            },
-        )
-        project_id = project_result.data["id"]
-
-        epic_result = await client.call_tool(
-            "createObject",
-            {
-                "kind": "epic",
-                "title": "Review Workflow Test Epic",
-                "projectRoot": planning_root,
-                "parent": project_id,
-            },
-        )
-        epic_id = epic_result.data["id"]
-
-        feature_result = await client.call_tool(
-            "createObject",
-            {
-                "kind": "feature",
-                "title": "Review Workflow Test Feature",
-                "projectRoot": planning_root,
-                "parent": epic_id,
-            },
-        )
-        feature_id = feature_result.data["id"]
-
-        # Step 2: Create mixed tasks with different priorities and timestamps
-        # Strategy: Create tasks in specific order to test timestamp ordering
-
-        # First, create standalone high priority task
-        standalone_high_result = await client.call_tool(
-            "createObject",
-            {
-                "kind": "task",
-                "title": "Standalone High Priority Review Task",
-                "projectRoot": planning_root,
-                "priority": "high",
-            },
-        )
-        standalone_high_id = standalone_high_result.data["id"]
-
-        # Short delay to ensure different timestamps
-        import asyncio
-
-        await asyncio.sleep(0.1)
-
-        # Create hierarchy normal priority task
-        hierarchy_normal_result = await client.call_tool(
-            "createObject",
-            {
-                "kind": "task",
-                "title": "Hierarchy Normal Priority Review Task",
-                "projectRoot": planning_root,
-                "parent": feature_id,
-                "priority": "normal",
-            },
-        )
-        hierarchy_normal_id = hierarchy_normal_result.data["id"]
-
-        await asyncio.sleep(0.1)
-
-        # Create standalone normal priority task
-        standalone_normal_result = await client.call_tool(
-            "createObject",
-            {
-                "kind": "task",
-                "title": "Standalone Normal Priority Review Task",
-                "projectRoot": planning_root,
-                "priority": "normal",
-            },
-        )
-        standalone_normal_id = standalone_normal_result.data["id"]
-
-        await asyncio.sleep(0.1)
-
-        # Create hierarchy high priority task (created last)
-        hierarchy_high_result = await client.call_tool(
-            "createObject",
-            {
-                "kind": "task",
-                "title": "Hierarchy High Priority Review Task",
-                "projectRoot": planning_root,
-                "parent": feature_id,
-                "priority": "high",
-            },
-        )
-        hierarchy_high_id = hierarchy_high_result.data["id"]
-
-        # Step 3: Transition all tasks to review status in reverse creation order
-        # This ensures different updated timestamps while maintaining creation order
-
-        # Update hierarchy high (created last, updated first)
-        await client.call_tool(
-            "updateObject",
-            {
-                "kind": "task",
-                "id": hierarchy_high_id,
-                "projectRoot": planning_root,
-                "yamlPatch": {"status": "in-progress"},
-            },
-        )
-        await client.call_tool(
-            "updateObject",
-            {
-                "kind": "task",
-                "id": hierarchy_high_id,
-                "projectRoot": planning_root,
-                "yamlPatch": {"status": "review"},
-            },
-        )
-
-        await asyncio.sleep(0.1)
-
-        # Update standalone normal (created third, updated second)
-        await client.call_tool(
-            "updateObject",
-            {
-                "kind": "task",
-                "id": standalone_normal_id,
-                "projectRoot": planning_root,
-                "yamlPatch": {"status": "in-progress"},
-            },
-        )
-        await client.call_tool(
-            "updateObject",
-            {
-                "kind": "task",
-                "id": standalone_normal_id,
-                "projectRoot": planning_root,
-                "yamlPatch": {"status": "review"},
-            },
-        )
-
-        await asyncio.sleep(0.1)
-
-        # Update hierarchy normal (created second, updated third)
-        await client.call_tool(
-            "updateObject",
-            {
-                "kind": "task",
-                "id": hierarchy_normal_id,
-                "projectRoot": planning_root,
-                "yamlPatch": {"status": "in-progress"},
-            },
-        )
-        await client.call_tool(
-            "updateObject",
-            {
-                "kind": "task",
-                "id": hierarchy_normal_id,
-                "projectRoot": planning_root,
-                "yamlPatch": {"status": "review"},
-            },
-        )
-
-        await asyncio.sleep(0.1)
-
-        # Update standalone high (created first, updated last)
-        await client.call_tool(
-            "updateObject",
-            {
-                "kind": "task",
-                "id": standalone_high_id,
-                "projectRoot": planning_root,
-                "yamlPatch": {"status": "in-progress"},
-            },
-        )
-        await client.call_tool(
-            "updateObject",
-            {
-                "kind": "task",
-                "id": standalone_high_id,
-                "projectRoot": planning_root,
-                "yamlPatch": {"status": "review"},
-            },
-        )
-
-        # Step 4: Test getNextReviewableTask returns oldest updated task
-        # Should return hierarchy_high_id (updated first, regardless of creation order)
-        review_result = await client.call_tool(
-            "getNextReviewableTask",
-            {"projectRoot": planning_root},
-        )
-
-        assert review_result.data is not None
-        assert review_result.data["task"] is not None
-        first_task = review_result.data["task"]
-        assert first_task["id"] == hierarchy_high_id
-        assert first_task["status"] == "review"
-        assert first_task["priority"] == "high"
-
-        # Step 5: Complete first task and verify next one is returned
-        await client.call_tool(
-            "completeTask",
-            {
-                "taskId": hierarchy_high_id,
-                "projectRoot": planning_root,
-            },
-        )
-
-        # Next should be standalone_normal_id (updated second)
-        review_result = await client.call_tool(
-            "getNextReviewableTask",
-            {"projectRoot": planning_root},
-        )
-
-        assert review_result.data is not None
-        assert review_result.data["task"] is not None
-        second_task = review_result.data["task"]
-        assert second_task["id"] == standalone_normal_id
-        assert second_task["status"] == "review"
-        assert second_task["priority"] == "normal"
-
-        # Step 6: Test that both standalone and hierarchy tasks appear in review queue
-        # Complete second task
-        await client.call_tool(
-            "completeTask",
-            {
-                "taskId": standalone_normal_id,
-                "projectRoot": planning_root,
-            },
-        )
-
-        # Next should be hierarchy_normal_id (updated third)
-        review_result = await client.call_tool(
-            "getNextReviewableTask",
-            {"projectRoot": planning_root},
-        )
-
-        assert review_result.data is not None
-        assert review_result.data["task"] is not None
-        third_task = review_result.data["task"]
-        assert third_task["id"] == hierarchy_normal_id
-        assert third_task["status"] == "review"
-        assert third_task["priority"] == "normal"
-
-        # Step 7: Complete third task and verify final standalone task
-        await client.call_tool(
-            "completeTask",
-            {
-                "taskId": hierarchy_normal_id,
-                "projectRoot": planning_root,
-            },
-        )
-
-        # Final should be standalone_high_id (updated last)
-        review_result = await client.call_tool(
-            "getNextReviewableTask",
-            {"projectRoot": planning_root},
-        )
-
-        assert review_result.data is not None
-        assert review_result.data["task"] is not None
-        final_task = review_result.data["task"]
-        assert final_task["id"] == standalone_high_id
-        assert final_task["status"] == "review"
-        assert final_task["priority"] == "high"
-
-        # Step 8: Complete final task and verify no more reviewable tasks
-        await client.call_tool(
-            "completeTask",
-            {
-                "taskId": standalone_high_id,
-                "projectRoot": planning_root,
-            },
-        )
-
-        # Should return None
-        review_result = await client.call_tool(
-            "getNextReviewableTask",
-            {"projectRoot": planning_root},
-        )
-
-        assert review_result.data is not None
-        assert review_result.data["task"] is None
-
-
-@pytest.mark.asyncio
-async def test_comprehensive_scope_filtering_mixed_environments(temp_dir):
-    """Test comprehensive scope filtering edge cases with mixed task environments.
-
-    Enhanced test covering project-level scope includes standalone tasks,
-    epic/feature-level scope combinations, and various edge cases.
-    """
-    # Create settings with temporary planning directory
-    settings = Settings(
-        planning_root=temp_dir / "planning",
-        debug_mode=True,
-        log_level="DEBUG",
-    )
-
-    # Create server instance
-    server = create_server(settings)
-    planning_root = str(temp_dir / "planning")
-
-    async with Client(server) as client:
-        # Step 1: Create multiple projects with hierarchical structure
-        project1_result = await client.call_tool(
-            "createObject",
-            {
-                "kind": "project",
-                "title": "Project 1 for Scope Testing",
-                "projectRoot": planning_root,
-            },
-        )
-        project1_id = project1_result.data["id"]
-
-        project2_result = await client.call_tool(
-            "createObject",
-            {
-                "kind": "project",
-                "title": "Project 2 for Scope Testing",
-                "projectRoot": planning_root,
-            },
-        )
-        project2_id = project2_result.data["id"]
-
-        # Create epics and features for project 1
-        epic1_result = await client.call_tool(
-            "createObject",
-            {
-                "kind": "epic",
-                "title": "Epic 1",
-                "projectRoot": planning_root,
-                "parent": project1_id,
-            },
-        )
-        epic1_id = epic1_result.data["id"]
-
-        feature1_result = await client.call_tool(
-            "createObject",
-            {
-                "kind": "feature",
-                "title": "Feature 1",
-                "projectRoot": planning_root,
-                "parent": epic1_id,
-            },
-        )
-        feature1_id = feature1_result.data["id"]
-
-        feature2_result = await client.call_tool(
-            "createObject",
-            {
-                "kind": "feature",
-                "title": "Feature 2",
-                "projectRoot": planning_root,
-                "parent": epic1_id,
-            },
-        )
-        feature2_id = feature2_result.data["id"]
-
-        # Create epic and feature for project 2
-        epic2_result = await client.call_tool(
-            "createObject",
-            {
-                "kind": "epic",
-                "title": "Epic 2",
-                "projectRoot": planning_root,
-                "parent": project2_id,
-            },
-        )
-        epic2_id = epic2_result.data["id"]
-
-        feature3_result = await client.call_tool(
-            "createObject",
-            {
-                "kind": "feature",
-                "title": "Feature 3",
-                "projectRoot": planning_root,
-                "parent": epic2_id,
-            },
-        )
-        feature3_id = feature3_result.data["id"]
-
-        # Step 2: Create comprehensive mixed task set
-        created_tasks = []
-
-        # Hierarchical tasks for project 1
-        for i, feature_id in enumerate([feature1_id, feature2_id]):
-            for j in range(2):  # 2 tasks per feature
-                task_result = await client.call_tool(
-                    "createObject",
-                    {
-                        "kind": "task",
-                        "title": f"Hierarchy Task {i + 1}-{j + 1}",
-                        "projectRoot": planning_root,
-                        "parent": feature_id,
-                        "priority": ["high", "normal"][j % 2],
-                    },
-                )
-                created_tasks.append(
-                    {
-                        "id": task_result.data["id"],
-                        "type": "hierarchy",
-                        "parent": feature_id,
-                        "project": project1_id,
-                        "epic": epic1_id,
-                    }
-                )
-
-        # Hierarchical tasks for project 2
-        for j in range(2):  # 2 tasks for project 2
-            task_result = await client.call_tool(
-                "createObject",
-                {
-                    "kind": "task",
-                    "title": f"Project 2 Hierarchy Task {j + 1}",
-                    "projectRoot": planning_root,
-                    "parent": feature3_id,
-                    "priority": ["normal", "low"][j % 2],
-                },
-            )
-            created_tasks.append(
-                {
-                    "id": task_result.data["id"],
-                    "type": "hierarchy",
-                    "parent": feature3_id,
-                    "project": project2_id,
-                    "epic": epic2_id,
-                }
-            )
-
-        # Standalone tasks (should be included in all project scopes)
-        for i in range(4):  # 4 standalone tasks
-            task_result = await client.call_tool(
-                "createObject",
-                {
-                    "kind": "task",
-                    "title": f"Standalone Task {i + 1}",
-                    "projectRoot": planning_root,
-                    "priority": ["high", "normal", "low", "normal"][i % 4],
-                },
-            )
-            created_tasks.append(
-                {
-                    "id": task_result.data["id"],
-                    "type": "standalone",
-                    "parent": None,
-                    "project": None,  # Standalone tasks don't belong to specific projects
-                    "epic": None,
-                }
-            )
-
-        # Step 3: Test project-level scope includes all standalone tasks
-        # Project 1 scope should include: 4 hierarchy tasks + 4 standalone tasks = 8 total
-        project1_result = await client.call_tool(
-            "listBacklog",
-            {"projectRoot": planning_root, "scope": project1_id},
-        )
-        assert project1_result.structured_content is not None
-        project1_tasks = project1_result.structured_content["tasks"]
-
-        # Should include all hierarchy tasks from project 1 AND all standalone tasks
-        assert len(project1_tasks) == 8  # 4 hierarchy + 4 standalone
-
-        # Verify standalone tasks are included in project scope
-        standalone_task_ids = {task["id"] for task in created_tasks if task["type"] == "standalone"}
-        project1_task_ids = {task["id"] for task in project1_tasks}
-
-        # All standalone tasks should be in project1 scope
-        assert standalone_task_ids.issubset(project1_task_ids)
-
-        # Verify hierarchy tasks from project 1 are included
-        project1_hierarchy_ids = {
-            task["id"]
-            for task in created_tasks
-            if task["type"] == "hierarchy" and task["project"] == project1_id
-        }
-        assert project1_hierarchy_ids.issubset(project1_task_ids)
-
-        # Verify hierarchy tasks from project 2 are NOT included
-        project2_hierarchy_ids = {
-            task["id"]
-            for task in created_tasks
-            if task["type"] == "hierarchy" and task["project"] == project2_id
-        }
-        assert project2_hierarchy_ids.isdisjoint(project1_task_ids)
-
-        # Step 4: Test project 2 scope also includes all standalone tasks
-        project2_result = await client.call_tool(
-            "listBacklog",
-            {"projectRoot": planning_root, "scope": project2_id},
-        )
-        assert project2_result.structured_content is not None
-        project2_tasks = project2_result.structured_content["tasks"]
-
-        # Should include: 2 hierarchy tasks from project 2 + 4 standalone tasks = 6 total
-        assert len(project2_tasks) == 6  # 2 hierarchy + 4 standalone
-
-        project2_task_ids = {task["id"] for task in project2_tasks}
-
-        # All standalone tasks should be in project2 scope too
-        assert standalone_task_ids.issubset(project2_task_ids)
-
-        # Verify hierarchy tasks from project 2 are included
-        assert project2_hierarchy_ids.issubset(project2_task_ids)
-
-        # Verify hierarchy tasks from project 1 are NOT included
-        assert project1_hierarchy_ids.isdisjoint(project2_task_ids)
-
-        # Step 5: Test epic-level scope filtering
-        # Epic 1 scope should include hierarchy tasks from features 1 & 2, but NO standalone tasks
-        epic1_result = await client.call_tool(
-            "listBacklog",
-            {"projectRoot": planning_root, "scope": epic1_id},
-        )
-        assert epic1_result.structured_content is not None
-        epic1_tasks = epic1_result.structured_content["tasks"]
-
-        # Should include only hierarchy tasks from epic1 (4 tasks from 2 features)
-        assert len(epic1_tasks) == 4
-
-        epic1_task_ids = {task["id"] for task in epic1_tasks}
-
-        # Should include all hierarchy tasks from epic1
-        epic1_hierarchy_ids = {
-            task["id"]
-            for task in created_tasks
-            if task["type"] == "hierarchy" and task["epic"] == epic1_id
-        }
-        assert epic1_hierarchy_ids == epic1_task_ids
-
-        # Should NOT include standalone tasks
-        assert standalone_task_ids.isdisjoint(epic1_task_ids)
-
-        # Step 6: Test feature-level scope filtering
-        # Feature 1 scope should include only its direct hierarchy tasks
-        feature1_result = await client.call_tool(
-            "listBacklog",
-            {"projectRoot": planning_root, "scope": feature1_id},
-        )
-        assert feature1_result.structured_content is not None
-        feature1_tasks = feature1_result.structured_content["tasks"]
-
-        # Should include only tasks directly under feature1 (2 tasks)
-        assert len(feature1_tasks) == 2
-
-        feature1_task_ids = {task["id"] for task in feature1_tasks}
-
-        # Should include only hierarchy tasks with feature1 as parent
-        feature1_hierarchy_ids = {
-            task["id"]
-            for task in created_tasks
-            if task["type"] == "hierarchy" and task["parent"] == feature1_id
-        }
-        assert feature1_hierarchy_ids == feature1_task_ids
-
-        # Should NOT include standalone tasks or tasks from other features
-        assert standalone_task_ids.isdisjoint(feature1_task_ids)
-
-        feature2_hierarchy_ids = {
-            task["id"]
-            for task in created_tasks
-            if task["type"] == "hierarchy" and task["parent"] == feature2_id
-        }
-        assert feature2_hierarchy_ids.isdisjoint(feature1_task_ids)
-
-        # Step 7: Test edge case - no scope filter (global)
-        # Should return ALL tasks (6 hierarchy + 4 standalone = 10 total)
-        global_result = await client.call_tool(
-            "listBacklog",
-            {"projectRoot": planning_root},
-        )
-        assert global_result.structured_content is not None
-        global_tasks = global_result.structured_content["tasks"]
-
-        assert len(global_tasks) == 10  # All tasks
-
-        global_task_ids = {task["id"] for task in global_tasks}
-        all_created_task_ids = {task["id"] for task in created_tasks}
-        assert global_task_ids == all_created_task_ids
-
-        # Step 8: Test combined scope filtering with status and priority
-        # Project1 scope + high priority should include high priority hierarchy tasks +
-        # high priority standalone
-        project1_high_result = await client.call_tool(
-            "listBacklog",
-            {
-                "projectRoot": planning_root,
-                "scope": project1_id,
-                "priority": "high",
-            },
-        )
-        assert project1_high_result.structured_content is not None
-        project1_high_tasks = project1_high_result.structured_content["tasks"]
-
-        # Verify all returned tasks have high priority
-        for task in project1_high_tasks:
-            assert task["priority"] == "high"
-
-        # Note: We need to count based on actual priority assignments
-        # From our creation: standalone tasks have priorities [high, normal, low, normal]
-        # Hierarchy tasks in project1 have priorities [high, normal, high, normal]
-        # So we expect 1 + 2 = 3 high priority tasks in project1 scope
-        assert len(project1_high_tasks) >= 1  # At least 1 high priority task
-
-
-@pytest.mark.asyncio
-async def test_comprehensive_task_lifecycle_mixed_environments(temp_dir):
-    """Test complete task lifecycle across both standalone and hierarchical task types.
-
-    Tests creation, claiming, status transitions, and completion for both task types
-    to ensure consistency and proper file movement operations.
-    """
-    # Create settings with temporary planning directory
-    settings = Settings(
-        planning_root=temp_dir / "planning",
-        debug_mode=True,
-        log_level="DEBUG",
-    )
-
-    # Create server instance
-    server = create_server(settings)
-    planning_root = str(temp_dir / "planning")
-
-    async with Client(server) as client:
-        # Step 1: Create hierarchical structure for hierarchy tasks
-        project_result = await client.call_tool(
-            "createObject",
-            {
-                "kind": "project",
-                "title": "Task Lifecycle Test Project",
-                "projectRoot": planning_root,
-            },
-        )
-        project_id = project_result.data["id"]
-
-        epic_result = await client.call_tool(
-            "createObject",
-            {
-                "kind": "epic",
-                "title": "Task Lifecycle Test Epic",
-                "projectRoot": planning_root,
-                "parent": project_id,
-            },
-        )
-        epic_id = epic_result.data["id"]
-
-        feature_result = await client.call_tool(
-            "createObject",
-            {
-                "kind": "feature",
-                "title": "Task Lifecycle Test Feature",
-                "projectRoot": planning_root,
-                "parent": epic_id,
-            },
-        )
-        feature_id = feature_result.data["id"]
-
-        # Step 2: Create mixed task types for lifecycle testing
-        # Create standalone task
-        standalone_task_result = await client.call_tool(
-            "createObject",
-            {
-                "kind": "task",
-                "title": "Standalone Lifecycle Task",
-                "projectRoot": planning_root,
-                "priority": "high",
-            },
-        )
-        standalone_task_id = standalone_task_result.data["id"]
-
-        # Create hierarchy task
-        hierarchy_task_result = await client.call_tool(
-            "createObject",
-            {
-                "kind": "task",
-                "title": "Hierarchy Lifecycle Task",
-                "projectRoot": planning_root,
-                "parent": feature_id,
-                "priority": "normal",
-            },
-        )
-        hierarchy_task_id = hierarchy_task_result.data["id"]
-
-        # Step 3: Verify initial state - both tasks should be in open status
-        standalone_initial = await client.call_tool(
-            "getObject",
-            {
-                "kind": "task",
-                "id": standalone_task_id,
-                "projectRoot": planning_root,
-            },
-        )
-        assert standalone_initial.data["yaml"]["status"] == "open"
-        assert standalone_initial.data["yaml"]["priority"] == "high"
-        assert standalone_initial.data["yaml"].get("parent") is None
-
-        hierarchy_initial = await client.call_tool(
-            "getObject",
-            {
-                "kind": "task",
-                "id": hierarchy_task_id,
-                "projectRoot": planning_root,
-            },
-        )
-        assert hierarchy_initial.data["yaml"]["status"] == "open"
-        assert hierarchy_initial.data["yaml"]["priority"] == "normal"
-        assert hierarchy_initial.data["yaml"]["parent"] == feature_id
-
-        # Step 4: Test claiming - standalone task should be claimed first (higher priority)
-        claim_result = await client.call_tool(
-            "claimNextTask",
-            {"projectRoot": planning_root, "worktree": "test-workspace-1"},
-        )
-
-        claimed_task = claim_result.data["task"]
-        assert claimed_task["id"] == standalone_task_id
-        assert claimed_task["status"] == "in-progress"
-        assert claimed_task["priority"] == "high"
-        assert claim_result.data["worktree"] == "test-workspace-1"
-
-        # Step 5: Verify standalone task status updated to in-progress
-        standalone_claimed = await client.call_tool(
-            "getObject",
-            {
-                "kind": "task",
-                "id": standalone_task_id,
-                "projectRoot": planning_root,
-            },
-        )
-        assert standalone_claimed.data["yaml"]["status"] == "in-progress"
-        assert "worktree" in standalone_claimed.data["yaml"]
-
-        # Step 6: Claim hierarchy task
-        claim_result2 = await client.call_tool(
-            "claimNextTask",
-            {"projectRoot": planning_root, "worktree": "test-workspace-2"},
-        )
-
-        claimed_task2 = claim_result2.data["task"]
-        assert claimed_task2["id"] == hierarchy_task_id
-        assert claimed_task2["status"] == "in-progress"
-        assert claimed_task2["priority"] == "normal"
-        assert claim_result2.data["worktree"] == "test-workspace-2"
-
-        # Step 7: Test status transitions - move both tasks to review
-        # Update standalone task to review
-        await client.call_tool(
-            "updateObject",
-            {
-                "kind": "task",
-                "id": standalone_task_id,
-                "projectRoot": planning_root,
-                "yamlPatch": {"status": "review"},
-            },
-        )
-
-        # Update hierarchy task to review
-        await client.call_tool(
-            "updateObject",
-            {
-                "kind": "task",
-                "id": hierarchy_task_id,
-                "projectRoot": planning_root,
-                "yamlPatch": {"status": "review"},
-            },
-        )
-
-        # Step 8: Verify both tasks in review status
-        standalone_review = await client.call_tool(
-            "getObject",
-            {
-                "kind": "task",
-                "id": standalone_task_id,
-                "projectRoot": planning_root,
-            },
-        )
-        assert standalone_review.data["yaml"]["status"] == "review"
-
-        hierarchy_review = await client.call_tool(
-            "getObject",
-            {
-                "kind": "task",
-                "id": hierarchy_task_id,
-                "projectRoot": planning_root,
-            },
-        )
-        assert hierarchy_review.data["yaml"]["status"] == "review"
-
-        # Step 9: Test getNextReviewableTask with mixed tasks
-        reviewable_result = await client.call_tool(
-            "getNextReviewableTask",
-            {"projectRoot": planning_root},
-        )
-
-        assert reviewable_result.data is not None
-        assert reviewable_result.data["task"] is not None
-        reviewable_task = reviewable_result.data["task"]
-
-        # Should return one of the review tasks (oldest updated timestamp)
-        assert reviewable_task["id"] in [standalone_task_id, hierarchy_task_id]
-        assert reviewable_task["status"] == "review"
-
-        # Step 10: Complete standalone task
-        complete_result = await client.call_tool(
-            "completeTask",
-            {
-                "taskId": standalone_task_id,
-                "projectRoot": planning_root,
-                "summary": "Completed standalone task lifecycle test",
-                "filesChanged": ["standalone_test.py", "config.yaml"],
-            },
-        )
-
-        assert complete_result.data is not None
-
-        # Step 11: Verify standalone task moved to done status
-        standalone_completed = await client.call_tool(
-            "getObject",
-            {
-                "kind": "task",
-                "id": standalone_task_id,
-                "projectRoot": planning_root,
-            },
-        )
-        assert standalone_completed.data["yaml"]["status"] == "done"
-
-        # Note: File movement to tasks-done is handled by completeTask internally
-        # The important verification is that the task status is updated to "done"
-
-        # Step 12: Complete hierarchy task
-        complete_result2 = await client.call_tool(
-            "completeTask",
-            {
-                "taskId": hierarchy_task_id,
-                "projectRoot": planning_root,
-                "summary": "Completed hierarchy task lifecycle test",
-                "filesChanged": ["hierarchy_test.py"],
-            },
-        )
-
-        assert complete_result2.data is not None
-
-        # Step 13: Verify hierarchy task moved to done status
-        hierarchy_completed = await client.call_tool(
-            "getObject",
-            {
-                "kind": "task",
-                "id": hierarchy_task_id,
-                "projectRoot": planning_root,
-            },
-        )
-        assert hierarchy_completed.data["yaml"]["status"] == "done"
-
-        # Note: File movement to tasks-done is handled by completeTask internally
-        # The important verification is that the task status is updated to "done"
-
-        # Step 14: Verify both completed tasks still appear in listBacklog
-        final_backlog = await client.call_tool(
-            "listBacklog",
-            {"projectRoot": planning_root},
-        )
-        assert final_backlog.structured_content is not None
-        final_tasks = final_backlog.structured_content["tasks"]
-
-        assert len(final_tasks) == 2
-        final_task_ids = {task["id"] for task in final_tasks}
-        assert final_task_ids == {standalone_task_id, hierarchy_task_id}
-
-        # Verify both tasks have done status
-        for task in final_tasks:
-            assert task["status"] == "done"
-
-        # Step 15: Verify no more reviewable tasks
-        no_review_result = await client.call_tool(
-            "getNextReviewableTask",
-            {"projectRoot": planning_root},
-        )
-        assert no_review_result.data["task"] is None
-
-        # Step 16: Verify no more claimable tasks
-        with pytest.raises(Exception) as exc_info:
-            await client.call_tool(
-                "claimNextTask",
-                {"projectRoot": planning_root},
-            )
-        assert "no" in str(exc_info.value).lower() and "available" in str(exc_info.value).lower()
-
-
-@pytest.mark.asyncio
-async def test_comprehensive_integration_all_mcp_operations(temp_dir):
-    """Test all MCP operations work consistently with mixed task environments.
-
-    End-to-end integration test covering all MCP tools with both standalone
-    and hierarchical tasks to ensure complete system integration.
-    """
-    # Create settings with temporary planning directory
-    settings = Settings(
-        planning_root=temp_dir / "planning",
-        debug_mode=True,
-        log_level="DEBUG",
-    )
-
-    # Create server instance
-    server = create_server(settings)
-    planning_root = str(temp_dir / "planning")
-
-    async with Client(server) as client:
-        # Step 1: Test health_check works
-        health_result = await client.call_tool("health_check")
-        assert health_result.data["status"] == "healthy"
-        assert health_result.data["planning_root"] == planning_root
-
-        # Step 2: Create comprehensive mixed environment using createObject
-        # Create project hierarchy
-        project_result = await client.call_tool(
-            "createObject",
-            {
-                "kind": "project",
-                "title": "Integration Test Project",
-                "projectRoot": planning_root,
-            },
-        )
-        project_id = project_result.data["id"]
-
-        epic_result = await client.call_tool(
-            "createObject",
-            {
-                "kind": "epic",
-                "title": "Integration Test Epic",
-                "projectRoot": planning_root,
-                "parent": project_id,
-            },
-        )
-        epic_id = epic_result.data["id"]
-
-        feature_result = await client.call_tool(
-            "createObject",
-            {
-                "kind": "feature",
-                "title": "Integration Test Feature",
-                "projectRoot": planning_root,
-                "parent": epic_id,
-            },
-        )
-        feature_id = feature_result.data["id"]
-
-        # Create mixed tasks
-        tasks_data = []
-
-        # Create hierarchy tasks with different priorities and statuses
-        for i in range(3):
-            task_result = await client.call_tool(
-                "createObject",
-                {
-                    "kind": "task",
-                    "title": f"Hierarchy Task {i + 1}",
-                    "projectRoot": planning_root,
-                    "parent": feature_id,
-                    "priority": ["high", "normal", "low"][i],
-                },
-            )
-            tasks_data.append(
-                {
-                    "id": task_result.data["id"],
-                    "type": "hierarchy",
-                    "priority": ["high", "normal", "low"][i],
-                }
-            )
-
-        # Create standalone tasks
-        for i in range(3):
-            task_result = await client.call_tool(
-                "createObject",
-                {
-                    "kind": "task",
-                    "title": f"Standalone Task {i + 1}",
-                    "projectRoot": planning_root,
-                    "priority": ["normal", "high", "low"][i],
-                },
-            )
-            tasks_data.append(
-                {
-                    "id": task_result.data["id"],
-                    "type": "standalone",
-                    "priority": ["normal", "high", "low"][i],
-                }
-            )
-
-        # Step 3: Test getObject for all created objects
-        # Test project retrieval
-        project_retrieved = await client.call_tool(
-            "getObject",
-            {
-                "kind": "project",
-                "id": project_id,
-                "projectRoot": planning_root,
-            },
-        )
-        assert project_retrieved.data["yaml"]["title"] == "Integration Test Project"
-
-        # Test epic retrieval
-        epic_retrieved = await client.call_tool(
-            "getObject",
-            {
-                "kind": "epic",
-                "id": epic_id,
-                "projectRoot": planning_root,
-            },
-        )
-        assert epic_retrieved.data["yaml"]["title"] == "Integration Test Epic"
-
-        # Test feature retrieval
-        feature_retrieved = await client.call_tool(
-            "getObject",
-            {
-                "kind": "feature",
-                "id": feature_id,
-                "projectRoot": planning_root,
-            },
-        )
-        assert feature_retrieved.data["yaml"]["title"] == "Integration Test Feature"
-
-        # Test task retrieval for both types
-        for task_data in tasks_data:
-            task_retrieved = await client.call_tool(
-                "getObject",
-                {
-                    "kind": "task",
-                    "id": task_data["id"],
-                    "projectRoot": planning_root,
-                },
-            )
-            assert task_retrieved.data["yaml"]["priority"] == task_data["priority"]
-
-        # Step 4: Test listBacklog with various filters
-        # Test global scope
-        global_backlog = await client.call_tool(
-            "listBacklog",
-            {"projectRoot": planning_root},
-        )
-        assert global_backlog.structured_content is not None
-        global_tasks = global_backlog.structured_content["tasks"]
-        assert len(global_tasks) == 6  # 3 hierarchy + 3 standalone
-
-        # Test project scope
-        project_backlog = await client.call_tool(
-            "listBacklog",
-            {"projectRoot": planning_root, "scope": project_id},
-        )
-        assert project_backlog.structured_content is not None
-        project_tasks = project_backlog.structured_content["tasks"]
-        assert len(project_tasks) == 6  # 3 hierarchy + 3 standalone (project scope includes all)
-
-        # Test feature scope
-        feature_backlog = await client.call_tool(
-            "listBacklog",
-            {"projectRoot": planning_root, "scope": feature_id},
-        )
-        assert feature_backlog.structured_content is not None
-        feature_tasks = feature_backlog.structured_content["tasks"]
-        assert len(feature_tasks) == 3  # Only hierarchy tasks under feature
-
-        # Test priority filtering
-        high_priority_backlog = await client.call_tool(
-            "listBacklog",
-            {"projectRoot": planning_root, "priority": "high"},
-        )
-        assert high_priority_backlog.structured_content is not None
-        high_priority_tasks = high_priority_backlog.structured_content["tasks"]
-        assert len(high_priority_tasks) == 2  # 1 hierarchy + 1 standalone with high priority
-
-        # Step 5: Test claimNextTask respects priority across both task types
-        claim_result = await client.call_tool(
-            "claimNextTask",
-            {"projectRoot": planning_root, "worktree": "integration-test"},
-        )
-
-        claimed_task = claim_result.data["task"]
-        assert claimed_task["priority"] == "high"  # Should claim highest priority first
-        assert claimed_task["status"] == "in-progress"
-        assert claim_result.data["worktree"] == "integration-test"
-
-        first_claimed_id = claimed_task["id"]
-
-        # Step 6: Test updateObject for both task types
-        # Find a standalone task and hierarchy task that aren't claimed
-        remaining_tasks = [t for t in tasks_data if t["id"] != first_claimed_id]
-        standalone_unclaimed = next(t for t in remaining_tasks if t["type"] == "standalone")
-        hierarchy_unclaimed = next(t for t in remaining_tasks if t["type"] == "hierarchy")
-
-        # Update standalone task priority
-        await client.call_tool(
-            "updateObject",
-            {
-                "kind": "task",
-                "id": standalone_unclaimed["id"],
-                "projectRoot": planning_root,
-                "yamlPatch": {"priority": "high"},
-            },
-        )
-
-        # Update hierarchy task priority
-        await client.call_tool(
-            "updateObject",
-            {
-                "kind": "task",
-                "id": hierarchy_unclaimed["id"],
-                "projectRoot": planning_root,
-                "yamlPatch": {"priority": "high"},
-            },
-        )
-
-        # Verify updates worked
-        standalone_updated = await client.call_tool(
-            "getObject",
-            {
-                "kind": "task",
-                "id": standalone_unclaimed["id"],
-                "projectRoot": planning_root,
-            },
-        )
-        assert standalone_updated.data["yaml"]["priority"] == "high"
-
-        hierarchy_updated = await client.call_tool(
-            "getObject",
-            {
-                "kind": "task",
-                "id": hierarchy_unclaimed["id"],
-                "projectRoot": planning_root,
-            },
-        )
-        assert hierarchy_updated.data["yaml"]["priority"] == "high"
-
-        # Step 7: Test review workflow
-        # Move claimed task to review
-        await client.call_tool(
-            "updateObject",
-            {
-                "kind": "task",
-                "id": first_claimed_id,
-                "projectRoot": planning_root,
-                "yamlPatch": {"status": "review"},
-            },
-        )
-
-        # Test getNextReviewableTask
-        reviewable_result = await client.call_tool(
-            "getNextReviewableTask",
-            {"projectRoot": planning_root},
-        )
-
-        assert reviewable_result.data is not None
-        assert reviewable_result.data["task"] is not None
-        reviewable_task = reviewable_result.data["task"]
-        assert reviewable_task["id"] == first_claimed_id
-        assert reviewable_task["status"] == "review"
-
-        # Step 8: Test completeTask
-        complete_result = await client.call_tool(
-            "completeTask",
-            {
-                "taskId": first_claimed_id,
-                "projectRoot": planning_root,
-                "summary": "Integration test task completed successfully",
-                "filesChanged": ["integration_test.py", "test_config.yaml"],
-            },
-        )
-
-        assert complete_result.data is not None
-
-        # Verify task marked as done
-        completed_task = await client.call_tool(
-            "getObject",
-            {
-                "kind": "task",
-                "id": first_claimed_id,
-                "projectRoot": planning_root,
-            },
-        )
-        assert completed_task.data["yaml"]["status"] == "done"
-
-        # Step 9: Test system state consistency after all operations
-        # Verify final listBacklog reflects all changes
-        final_backlog = await client.call_tool(
-            "listBacklog",
-            {"projectRoot": planning_root},
-        )
-        assert final_backlog.structured_content is not None
-        final_tasks = final_backlog.structured_content["tasks"]
-        assert len(final_tasks) == 6  # All tasks still visible
-
-        # Verify priority updates are reflected
-        updated_priorities = {task["id"]: task["priority"] for task in final_tasks}
-        assert updated_priorities[standalone_unclaimed["id"]] == "high"
-        assert updated_priorities[hierarchy_unclaimed["id"]] == "high"
-        assert updated_priorities[first_claimed_id] == "high"
-
-        # Verify status update is reflected
-        task_statuses = {task["id"]: task["status"] for task in final_tasks}
-        assert task_statuses[first_claimed_id] == "done"
-
-        # Step 10: Test that no more reviewable tasks exist
-        no_review_result = await client.call_tool(
-            "getNextReviewableTask",
-            {"projectRoot": planning_root},
-        )
-        assert no_review_result.data["task"] is None
-
-        # Step 11: Test health check still works after all operations
-        final_health_result = await client.call_tool("health_check")
-        assert final_health_result.data["status"] == "healthy"
